requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.0
pdfplumber>=0.10.0
reportlab>=4.0
gunicorn>=21.0
whitenoise>=6.6
dj-database-url>=2.1
psycopg2-binary>=2.9
# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
regex>=2023.0.0
# Optional: AI normalization (lightweight)
openai>=1.0.0